requirements.txt
apache-airflow
pandas
pytest
pytest-mock
python-dotenv
requests
supabase
streamlit
